Fix tests in Go 1.7/1.8
Go 1.7/1.8 seems to run more tests than necessary when executing `go
test ./...` and tries to compile much more than necessary.
When running `dep ensure` it seems to happen that
the `spf13/cobra/cobra` directory is still present and (probably due to
a bug) `go test` will try to compile this in. In the dep package list we
don't have some of the dependencies the cobra CLI uses, therefore the
test compilation fails.